requirements.txt
openai
google-cloud-speech
google-generativeai
dotenv